Reorganized components according to convention
