+++ app/components/document-creator.js
@@ -1,28 +1,3 @@
-import Ember from 'ember';
+import DocumentCreator from 'ember-gdrive/components/document-creator';
 
-export default Ember.Component.extend({
-    tagName: 'document-creator',
-    classNames: 'document-creator',
-    documentTitle: null,
-    isCreating: false,
-
-    actions: {
-      createDocument: function () {
-        this.createDocument();
-      },
-    },
-
-    createDocument: function () {
-      var title = this.get('documentTitle').trim(),
-        component = this;
-
-      this.set('isCreating', true);
-
-      return this.get('documentSource').createDocument(title).then(function (doc) {
-        component.sendAction('documentCreated', doc);
-        component.set('isCreating', false);
-      }, function (error) {
-        component.set('isCreating', false);
-      });
-    }
-});
+export default DocumentCreator;
